+++ next-docs/public/examples/authCode/FourChars.tsx
@@ -6,25 +6,21 @@ const Example = () => {
   const { control, handleSubmit } = useForm({
     defaultValues: {
       authCode: '',
-    }
+    },
   });
-
   const codeLength = 4;
-
   const onSubmit = ({ authCode }: { authCode: string }) => {
-    if (authCode.length === codeLength)
-      alert(`Code sent: "${authCode}"`);
+    if (authCode.length === codeLength) alert(`Code sent: "${authCode}"`);
   };
-
   return (
     <Form
       onSubmit={handleSubmit(onSubmit)}
-      className='flex flex-col justify-between items-end gap-y-2'
+      className="flex flex-col justify-between items-end gap-y-2"
     >
       <Controller
-        name='authCode'
+        name="authCode"
         control={control}
-        render={({ field }) =>
+        render={({ field }) => (
           <AuthCode
             {...field}
             length={codeLength}
@@ -34,10 +30,10 @@ const Example = () => {
             }}
             disabled={field.value.length === codeLength}
           />
-        }
+        )}
       />
     </Form>
   );
-}
+};
 
 export default Example;