+++ next-docs/tailwind.config.js
@@ -4,6 +4,7 @@ module.exports = {
     './pages/**/*.{js,ts,jsx,tsx}',
     './components/**/*.{js,ts,jsx,tsx}',
     './node_modules/@heathmont/moon-core-tw/**/*.{js,ts,jsx,tsx}',
+    './node_modules/@heathmont/moon-cmdk-tw/**/*.{js,ts,jsx,tsx}',
     './node_modules/@heathmont/moon-table-tw/**/*.{js,ts,jsx,tsx}',
   ],
   presets: [
@@ -11,17 +12,9 @@ module.exports = {
   ],
   theme: {
     extend: {
-      colors: {
-        active: {
-          DEFAULT: '#4e46b4',
-          80: 'rgba(78, 70, 180, 0.1)',
-          60: '#e9e7e9',
-        },
-      },
       screens: {
-        '3xl': '1800px',
+        '3xl': '2100px',
       },
     },
   },
-  plugins: [require('tailwindcss-rtl')],
 };