+++ workspaces/base/src/index.ts
@@ -8,6 +8,8 @@ export { default as Button } from './button/Button';
 export * from './button/Button';
 export { default as Checkbox } from './checkbox/Checkbox';
 export * from './checkbox/Checkbox';
+export { default as Chip } from './chip/Chip';
+export * from './chip/Chip';
 export { default as Input } from './input/Input';
 export * from './input/Input';
 export { default as Loader } from './loader/Loader';