+++ workspaces/core/src/tabs/Tabs.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import { Tab as HeadlesssTab } from '@headlessui/react';
-import mergeClassnames from '../mergeClassnames/mergeClassnames';
-import getTabSize from './private/utils/getTabSize';
-import TabsContext from './private/utils/TabsContext';
-import useTabsContext from './private/utils/useTabsContext';
 import type ListProps from './private/types/ListProps';
 import type PanelProps from './private/types/PanelProps';
 import type TabProps from './private/types/TabProps';
 import type TabsProps from './private/types/TabsProps';
+import getTabSize from './private/utils/getTabSize';
+import TabsContext from './private/utils/TabsContext';
+import useTabsContext from './private/utils/useTabsContext';
+import mergeClassnames from '../mergeClassnames/mergeClassnames';
 
 const TabsRoot = ({ children, selectedIndex, onChange }: TabsProps) => (
   <HeadlesssTab.Group selectedIndex={selectedIndex} onChange={onChange}>
@@ -62,13 +62,15 @@ const Tab = React.forwardRef(
           mergeClassnames(
             getTabSize(size),
             'relative flex items-center justify-center w-full whitespace-nowrap text-moon-14',
-            'text-bulma font-medium cursor-pointer after:content-[""] after:absolute after:left-0',
-            'after:bottom-0 after:w-full after:h-[2px] after:bg-piccolo after:transition-transform',
-            'after:duration-300 after:origin-top-left after:scale-x-0 after:scale-y-100',
-            'hover:after:origin-top-left hover:after:scale-100 hover:text-piccolo',
+            'text-bulma font-medium cursor-pointer after:absolute after:start-0 after:bottom-0',
+            'after:w-full after:h-[2px] after:bg-piccolo after:transition-transform',
+            'after:duration-300 ltr:after:origin-top-left rtl:after:origin-top-right',
+            'after:scale-x-0 after:scale-y-100 ltr:hover:after:origin-top-left',
+            'rtl:hover:after:origin-top-right hover:after:scale-100 hover:text-piccolo',
             'focus:outline-none',
             isDisabled && 'opacity-60 hover:after:scale-0 cursor-not-allowed',
-            selected && 'after:origin-top-left after:scale-x-100 text-piccolo',
+            selected &&
+              'ltr:after:origin-top-left rtl:after:origin-top-right after:scale-x-100 text-piccolo',
             typeof className === 'function'
               ? className({ selected: selected })
               : className