+++ docs/CHANGELOG.md
@@ -1,5 +1,86 @@
 # docs
 
+## 0.5.2
+
+### Patch Changes
+
+- Deploy 10.7.2
+- Updated dependencies
+  - @heathmont/moon-core-tw@10.7.2
+  - @heathmont/moon-base-tw@10.7.2
+
+## 0.5.1
+
+### Patch Changes
+
+- Deploy 10.7.1
+- Updated dependencies
+  - @heathmont/moon-base-tw@10.7.1
+  - @heathmont/moon-core-tw@10.7.1
+
+## 0.5.0
+
+### Minor Changes
+
+- Deploy 10.7.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @heathmont/moon-core-tw@10.7.0
+  - @heathmont/moon-base-tw@10.7.0
+
+## 0.4.3
+
+### Patch Changes
+
+- Deploy 10.6.3
+- Updated dependencies
+  - @heathmont/moon-base-tw@10.6.3
+  - @heathmont/moon-core-tw@10.6.3
+
+## 0.4.2
+
+### Patch Changes
+
+- Deploy 10.6.2
+- Updated dependencies
+  - @heathmont/moon-base-tw@10.6.2
+  - @heathmont/moon-core-tw@10.6.2
+
+## 0.4.1
+
+### Patch Changes
+
+- Deploy
+- Updated dependencies
+  - @heathmont/moon-base-tw@10.6.1
+  - @heathmont/moon-core-tw@10.6.1
+
+## 0.4.0
+
+### Minor Changes
+
+- Deploy 10.6.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @heathmont/moon-core-tw@10.6.0
+  - @heathmont/moon-base-tw@10.6.0
+
+## 0.3.0
+
+### Minor Changes
+
+- Bump version to 10.5.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @heathmont/moon-base-tw@10.5.0
+  - @heathmont/moon-core-tw@10.5.0
+
 ## 0.2.0
 
 ### Minor Changes