+++ next-docs/CHANGELOG.md
@@ -1,5 +1,206 @@
 # Change Log
 
+## 10.7.2
+
+### Patch Changes
+
+- Deploy 10.7.2
+- Updated dependencies
+  - @heathmont/moon-themes-tw@10.7.2
+  - @heathmont/moon-core-tw@10.7.2
+  - @heathmont/moon-cmdk-tw@10.7.2
+  - @heathmont/moon-table-tw@10.7.2
+
+## 10.7.1
+
+### Patch Changes
+
+- Deploy 10.7.1
+- Updated dependencies
+  - @heathmont/moon-accordion@10.7.1
+  - @heathmont/moon-assets@10.7.1
+  - @heathmont/moon-charts@10.7.1
+  - @heathmont/moon-components@10.7.1
+  - @heathmont/moon-core@10.7.1
+  - @heathmont/moon-datepicker@10.7.1
+  - @heathmont/moon-draggabletable@10.7.1
+  - @heathmont/moon-icons@10.7.1
+  - @heathmont/moon-pagination@10.7.1
+  - @heathmont/moon-popover@10.7.1
+  - @heathmont/moon-select@10.7.1
+  - @heathmont/moon-sidebar@10.7.1
+  - @heathmont/moon-table@10.7.1
+  - @heathmont/moon-themes@10.7.1
+  - @heathmont/moon-utils@10.7.1
+  - @heathmont/moon-cmdk-tw@10.7.1
+  - @heathmont/moon-core-tw@10.7.1
+  - @heathmont/moon-table-tw@10.7.1
+  - @heathmont/moon-themes-tw@10.7.1
+
+## 10.7.0
+
+### Minor Changes
+
+- Deploy 10.7.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @heathmont/moon-core-tw@10.7.0
+  - @heathmont/moon-accordion@10.7.0
+  - @heathmont/moon-assets@10.7.0
+  - @heathmont/moon-charts@10.7.0
+  - @heathmont/moon-components@10.7.0
+  - @heathmont/moon-core@10.7.0
+  - @heathmont/moon-datepicker@10.7.0
+  - @heathmont/moon-draggabletable@10.7.0
+  - @heathmont/moon-icons@10.7.0
+  - @heathmont/moon-pagination@10.7.0
+  - @heathmont/moon-popover@10.7.0
+  - @heathmont/moon-select@10.7.0
+  - @heathmont/moon-sidebar@10.7.0
+  - @heathmont/moon-table@10.7.0
+  - @heathmont/moon-themes@10.7.0
+  - @heathmont/moon-utils@10.7.0
+  - @heathmont/moon-cmdk-tw@10.7.0
+  - @heathmont/moon-table-tw@10.7.0
+  - @heathmont/moon-themes-tw@10.7.0
+
+## 10.6.3
+
+### Patch Changes
+
+- Deploy 10.6.3
+- Updated dependencies
+  - @heathmont/moon-accordion@10.6.3
+  - @heathmont/moon-assets@10.6.3
+  - @heathmont/moon-charts@10.6.3
+  - @heathmont/moon-components@10.6.3
+  - @heathmont/moon-core@10.6.3
+  - @heathmont/moon-datepicker@10.6.3
+  - @heathmont/moon-draggabletable@10.6.3
+  - @heathmont/moon-icons@10.6.3
+  - @heathmont/moon-pagination@10.6.3
+  - @heathmont/moon-popover@10.6.3
+  - @heathmont/moon-select@10.6.3
+  - @heathmont/moon-sidebar@10.6.3
+  - @heathmont/moon-table@10.6.3
+  - @heathmont/moon-themes@10.6.3
+  - @heathmont/moon-utils@10.6.3
+  - @heathmont/moon-cmdk-tw@10.6.3
+  - @heathmont/moon-core-tw@10.6.3
+  - @heathmont/moon-table-tw@10.6.3
+  - @heathmont/moon-themes-tw@10.6.3
+
+## 10.6.2
+
+### Patch Changes
+
+- Deploy 10.6.2
+- Updated dependencies
+  - @heathmont/moon-accordion@10.6.2
+  - @heathmont/moon-assets@10.6.2
+  - @heathmont/moon-charts@10.6.2
+  - @heathmont/moon-components@10.6.2
+  - @heathmont/moon-core@10.6.2
+  - @heathmont/moon-datepicker@10.6.2
+  - @heathmont/moon-draggabletable@10.6.2
+  - @heathmont/moon-icons@10.6.2
+  - @heathmont/moon-pagination@10.6.2
+  - @heathmont/moon-popover@10.6.2
+  - @heathmont/moon-select@10.6.2
+  - @heathmont/moon-sidebar@10.6.2
+  - @heathmont/moon-table@10.6.2
+  - @heathmont/moon-themes@10.6.2
+  - @heathmont/moon-utils@10.6.2
+  - @heathmont/moon-cmdk-tw@10.6.2
+  - @heathmont/moon-core-tw@10.6.2
+  - @heathmont/moon-table-tw@10.6.2
+  - @heathmont/moon-themes-tw@10.6.2
+
+## 10.6.1
+
+### Patch Changes
+
+- Deploy
+- Updated dependencies
+  - @heathmont/moon-cmdk-tw@10.6.1
+  - @heathmont/moon-accordion@10.6.1
+  - @heathmont/moon-assets@10.6.1
+  - @heathmont/moon-charts@10.6.1
+  - @heathmont/moon-components@10.6.1
+  - @heathmont/moon-core@10.6.1
+  - @heathmont/moon-datepicker@10.6.1
+  - @heathmont/moon-draggabletable@10.6.1
+  - @heathmont/moon-icons@10.6.1
+  - @heathmont/moon-pagination@10.6.1
+  - @heathmont/moon-popover@10.6.1
+  - @heathmont/moon-select@10.6.1
+  - @heathmont/moon-sidebar@10.6.1
+  - @heathmont/moon-table@10.6.1
+  - @heathmont/moon-themes@10.6.1
+  - @heathmont/moon-utils@10.6.1
+  - @heathmont/moon-core-tw@10.6.1
+  - @heathmont/moon-table-tw@10.6.1
+  - @heathmont/moon-themes-tw@10.6.1
+
+## 10.6.0
+
+### Minor Changes
+
+- Deploy 10.6.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @heathmont/moon-themes-tw@10.6.0
+  - @heathmont/moon-cmdk-tw@10.6.0
+  - @heathmont/moon-core-tw@10.6.0
+  - @heathmont/moon-accordion@10.6.0
+  - @heathmont/moon-assets@10.6.0
+  - @heathmont/moon-charts@10.6.0
+  - @heathmont/moon-components@10.6.0
+  - @heathmont/moon-core@10.6.0
+  - @heathmont/moon-datepicker@10.6.0
+  - @heathmont/moon-draggabletable@10.6.0
+  - @heathmont/moon-icons@10.6.0
+  - @heathmont/moon-pagination@10.6.0
+  - @heathmont/moon-popover@10.6.0
+  - @heathmont/moon-select@10.6.0
+  - @heathmont/moon-sidebar@10.6.0
+  - @heathmont/moon-table@10.6.0
+  - @heathmont/moon-themes@10.6.0
+  - @heathmont/moon-utils@10.6.0
+  - @heathmont/moon-table-tw@10.6.0
+
+## 10.5.0
+
+### Minor Changes
+
+- Bump version to 10.5.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @heathmont/moon-table-tw@10.5.0
+  - @heathmont/moon-themes-tw@10.5.0
+  - @heathmont/moon-assets@10.5.0
+  - @heathmont/moon-core-tw@10.5.0
+  - @heathmont/moon-accordion@10.5.0
+  - @heathmont/moon-charts@10.5.0
+  - @heathmont/moon-components@10.5.0
+  - @heathmont/moon-core@10.5.0
+  - @heathmont/moon-datepicker@10.5.0
+  - @heathmont/moon-draggabletable@10.5.0
+  - @heathmont/moon-icons@10.5.0
+  - @heathmont/moon-pagination@10.5.0
+  - @heathmont/moon-popover@10.5.0
+  - @heathmont/moon-select@10.5.0
+  - @heathmont/moon-sidebar@10.5.0
+  - @heathmont/moon-table@10.5.0
+  - @heathmont/moon-themes@10.5.0
+  - @heathmont/moon-utils@10.5.0
+
 ## 10.4.0
 
 ### Minor Changes