+++ packages/sidebar/CHANGELOG.md
@@ -1,5 +1,77 @@
 # Change Log
 
+## 10.7.1
+
+### Patch Changes
+
+- Deploy 10.7.1
+- Updated dependencies
+  - @heathmont/moon-themes@10.7.1
+  - @heathmont/moon-utils@10.7.1
+
+## 10.7.0
+
+### Minor Changes
+
+- Deploy 10.7.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @heathmont/moon-themes@10.7.0
+  - @heathmont/moon-utils@10.7.0
+
+## 10.6.3
+
+### Patch Changes
+
+- Deploy 10.6.3
+- Updated dependencies
+  - @heathmont/moon-themes@10.6.3
+  - @heathmont/moon-utils@10.6.3
+
+## 10.6.2
+
+### Patch Changes
+
+- Deploy 10.6.2
+- Updated dependencies
+  - @heathmont/moon-themes@10.6.2
+  - @heathmont/moon-utils@10.6.2
+
+## 10.6.1
+
+### Patch Changes
+
+- Deploy
+- Updated dependencies
+  - @heathmont/moon-themes@10.6.1
+  - @heathmont/moon-utils@10.6.1
+
+## 10.6.0
+
+### Minor Changes
+
+- Deploy 10.6.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @heathmont/moon-themes@10.6.0
+  - @heathmont/moon-utils@10.6.0
+
+## 10.5.0
+
+### Minor Changes
+
+- Bump version to 10.5.0
+
+### Patch Changes
+
+- Updated dependencies
+  - @heathmont/moon-themes@10.5.0
+  - @heathmont/moon-utils@10.5.0
+
 ## 10.4.0
 
 ### Minor Changes